+++ src/main/java/com/airbnl/managerservice/model/Reservation.java
@@ -9,10 +9,10 @@ import java.sql.Timestamp;
 @Getter
 @Setter
 public class Reservation {
-    private int id;
-    private int userId;
+    private long id;
+    private long userId;
     private String username;
-    private int roomId;
+    private long roomId;
     private String roomName;
     private double cost;
     private Timestamp dateFrom;
