+++ src/main/java/com/airbnl/managerservice/model/Room.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Room {
-    private int id;
-    private int hotelId;
+    private long id;
+    private long hotelId;
     private String number;
-    private int typeId;
+    private long typeId;
     private int bedCount;
 }
